Consolidate classes
